Delete old page component and new page modify
