+++ app/blog/[slug]/page.tsx
@@ -22,6 +22,8 @@ import {
   getAllTags,
   getAllBlocksByBlockId,
 } from '../../../lib/notion/client'
+import LikeButton from '../../../components/like-button'
+import TocLink from '../../../components/toc'
 
 export const revalidate = 30
 export const dynamicParams = false
@@ -77,6 +79,7 @@ const BlogSlugPage = async ({ params: { slug } }) => {
                 id={post.Slug}
               />
             )}
+            <LikeButton slug={post.Slug} like={post.Like} />
           </footer>
         </div>
       </div>
@@ -89,6 +92,7 @@ const BlogSlugPage = async ({ params: { slug } }) => {
         <BlogPostLink heading="Recommended" posts={rankedPosts} />
         <BlogPostLink heading="Latest posts" posts={recentPosts} />
         <BlogTagLink heading="Categories" tags={tags} />
+        <TocLink blocks={blocks}/>
       </div>
     </div>
   )
