+++ app/blog/before/[date]/page.tsx
@@ -1,4 +1,4 @@
-import { notFound } from 'next/navigation'
+import { notFound, useRouter } from 'next/navigation'
 import { NUMBER_OF_POSTS_PER_PAGE } from '../../../../app/server-constants'
 import {
   getRankedPosts,
@@ -17,12 +17,15 @@ import {
   PostTitle,
   ReadMoreLink,
 } from '../../../../components/blog-parts'
-import styles from '../../../../styles/blog.module.css'
+import styles from '../../../../styles/blog.module.scss'
+import Mystyles from '../../../../styles/mystyles.module.scss'
+import Masonry from 'react-masonry-css'
 
 export const revalidate = 3600
 
 const BlogBeforeDatePage = async ({ params: { date: encodedDate } }) => {
   const date = decodeURIComponent(encodedDate)
+  const router = useRouter()
 
   if (!Date.parse(date) || !/^\d{4}-\d{2}-\d{2}/.test(date)) {
     notFound()
@@ -35,6 +38,14 @@ const BlogBeforeDatePage = async ({ params: { date: encodedDate } }) => {
     getAllTags(),
   ])
 
+  const breakpointColumnsObj = {
+    default: 2,
+    1280: 2,
+    1100: 2,
+    800: 2,
+    500: 1
+  };
+
   return (
     <div className={styles.container}>
       <div className={styles.mainContent}>
@@ -43,6 +54,11 @@ const BlogBeforeDatePage = async ({ params: { date: encodedDate } }) => {
         </header>
 
         <NoContents contents={posts} />
+        <Masonry
+          breakpointCols={breakpointColumnsObj}
+          className="my-masonry-grid"
+          columnClassName="my-masonry-grid_column"
+        >
 
         {posts.map(post => {
           return (
@@ -55,9 +71,15 @@ const BlogBeforeDatePage = async ({ params: { date: encodedDate } }) => {
             </div>
           )
         })}
+        </Masonry>
 
         <footer>
-          <NextPageLink firstPost={firstPost} posts={posts} />
+          <div className={Mystyles.PageLinkContainer}>
+            <div>
+              <a onClick={() => router.back()}>← Newer Page</a>
+            </div>
+            <NextPageLink firstPost={firstPost} posts={posts} />
+          </div>
         </footer>
       </div>
