+++ app/blog/page.tsx
@@ -8,15 +8,16 @@ import {
   PostExcerpt,
   PostTags,
   PostTitle,
-  ReadMoreLink,
 } from '../../components/blog-parts'
-import styles from '../../styles/blog.module.css'
+import styles from '../../styles/blog.module.scss'
+import Mystyles from '../../styles/mystyles.module.scss'
 import {
   getPosts,
   getFirstPost,
   getRankedPosts,
   getAllTags,
 } from '../../lib/notion/client'
+import Masonry from 'react-masonry-css'
 
 export const revalidate = 60
 
@@ -28,10 +29,23 @@ const BlogPage = async () => {
     getAllTags(),
   ])
 
+  const breakpointColumnsObj = {
+    default: 2,
+    1280: 2,
+    1100: 2,
+    800: 2,
+    500: 1
+  };
+
   return (
     <div className={styles.container}>
       <div className={styles.mainContent}>
         <NoContents contents={posts} />
+        <Masonry
+          breakpointCols={breakpointColumnsObj}
+          className="my-masonry-grid"
+          columnClassName="my-masonry-grid_column"
+        >
 
         {posts.map(post => {
           return (
@@ -40,13 +54,16 @@ const BlogPage = async () => {
               <PostTags post={post} />
               <PostTitle post={post} />
               <PostExcerpt post={post} />
-              <ReadMoreLink post={post} />
+              {/* <ReadMoreLink post={post} /> */}
             </div>
           )
         })}
+        </Masonry>
 
         <footer>
-          <NextPageLink firstPost={firstPost} posts={posts} />
+          <div className={Mystyles.nextPageLink}>
+            <NextPageLink firstPost={firstPost} posts={posts} />
+          </div>
         </footer>
       </div>
