+++ app/blog/tag/[tag]/page.tsx
@@ -8,9 +8,8 @@ import {
   PostExcerpt,
   PostTags,
   PostTitle,
-  ReadMoreLink,
 } from '../../../../components/blog-parts'
-import styles from '../../../../styles/blog.module.css'
+import styles from '../../../../styles/blog.module.scss'
 import {
   getPosts,
   getRankedPosts,
@@ -18,6 +17,7 @@ import {
   getFirstPostByTag,
   getAllTags,
 } from '../../../../lib/notion/client'
+import Masonry from 'react-masonry-css'
 
 export const revalidate = 60
 export const dynamicParams = false
@@ -43,12 +43,25 @@ const BlogTagPage = async ({ params: { tag: encodedTag } }) => {
     getAllTags(),
   ])
 
+  const breakpointColumnsObj = {
+    default: 2,
+    1280: 2,
+    1100: 2,
+    800: 2,
+    500: 1
+  };
+
   return (
     <div className={styles.container}>
       <div className={styles.mainContent}>
         <header>
           <h2>{tag}</h2>
         </header>
+        <Masonry
+          breakpointCols={breakpointColumnsObj}
+          className="my-masonry-grid"
+          columnClassName="my-masonry-grid_column"
+        >
 
         {posts.map(post => {
           return (
@@ -57,10 +70,11 @@ const BlogTagPage = async ({ params: { tag: encodedTag } }) => {
               <PostTags post={post} />
               <PostTitle post={post} />
               <PostExcerpt post={post} />
-              <ReadMoreLink post={post} />
+              {/* <ReadMoreLink post={post} /> */}
             </div>
           )
         })}
+        </Masonry>
 
         <footer>
           <NextPageLink firstPost={firstPost} posts={posts} tag={tag} />
