+++ app/page.tsx
@@ -1,11 +1,8 @@
 import Link from 'next/link'
-import { NEXT_PUBLIC_SITE_TITLE } from './server-constants'
-import GoogleAnalytics from '../components/google-analytics'
-import styles from '../styles/page.module.css'
+import styles from '../styles/page.module.scss'
 
 const RootPage = () => (
   <>
-    <GoogleAnalytics pageTitle={NEXT_PUBLIC_SITE_TITLE} />
     <div className={styles.container}>
       <div>
         <h2>Welcome!</h2>
