--- pages/blog/before/[date].tsx
@@ -1,138 +1,0 @@
-import React, { useEffect } from 'react'
-import { useRouter } from 'next/router'
-
-import { NUMBER_OF_POSTS_PER_PAGE } from '../../../lib/notion/server-constants'
-import DocumentHead from '../../../components/document-head'
-import {
-  BlogPostLink,
-  BlogTagLink,
-  NextPageLink,
-  NoContents,
-  PostDate,
-  PostExcerpt,
-  PostTags,
-  PostTitle,
-  PostsNotFound,
-} from '../../../components/blog-parts'
-import styles from '../../../styles/blog.module.scss'
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-import Mystyles from '../../../styles/mystyles.module.scss'
-
-import { getBeforeLink } from '../../../lib/blog-helpers'
-import {
-  getPosts,
-  getRankedPosts,
-  getPostsBefore,
-  getFirstPost,
-  getAllTags,
-} from '../../../lib/notion/client'
-import Masonry from 'react-masonry-css'
-
-export async function getStaticProps({ params: { date } }) {
-  if (!Date.parse(date) || !/\d{4}-\d{2}-\d{2}/.test(date)) {
-    return { notFound: true }
-  }
-
-  const [posts, firstPost, rankedPosts, tags] = await Promise.all([
-    getPostsBefore(date, NUMBER_OF_POSTS_PER_PAGE),
-    getFirstPost(),
-    getRankedPosts(),
-    getAllTags(),
-  ])
-
-  return {
-    props: {
-      date,
-      posts,
-      firstPost,
-      rankedPosts,
-      tags,
-    },
-    revalidate: 3600,
-  }
-}
-
-export async function getStaticPaths() {
-  const posts = await getPosts()
-  const path = getBeforeLink(posts[posts.length - 1].Date)
-
-  return {
-    paths: [path],
-    fallback: 'blocking',
-  }
-}
-
-const RenderPostsBeforeDate = ({
-  date,
-  posts = [],
-  firstPost,
-  rankedPosts = [],
-  tags = [],
-  redirect,
-}) => {
-  const router = useRouter()
-
-  useEffect(() => {
-    if (redirect && !posts) {
-      router.replace(redirect)
-    }
-  }, [router, redirect, posts])
-
-  if (!posts) {
-    return <PostsNotFound />
-  }
-
-  const breakpointColumnsObj = {
-    default: 2,
-    1280: 2,
-    1100: 2,
-    800: 2,
-    500: 1
-  };
-  return (
-    <div className={styles.container}>
-      <DocumentHead description={`Post before ${date.split('T')[0]}`} />
-
-      <div className={styles.mainContent}>
-        <header>
-          <h2>Posts before {date.split('T')[0]}</h2>
-        </header>
-
-        <NoContents contents={posts} />
-        <Masonry
-          breakpointCols={breakpointColumnsObj}
-          className="my-masonry-grid"
-          columnClassName="my-masonry-grid_column"
-        >
-
-        {posts.map(post => {
-          return (
-            <div className={styles.post} key={post.Slug}>
-              <PostDate post={post} />
-              <PostTags post={post} />
-              <PostTitle post={post} />
-              <PostExcerpt post={post} />
-            </div>
-          )
-        })}
-        </Masonry>
-
-        <footer>
-          <div className={Mystyles.PageLinkContainer}>
-            <div>
-              <a onClick={() => router.back()}>← Newer Page</a>
-            </div>
-            <NextPageLink firstPost={firstPost} posts={posts} />
-          </div>
-        </footer>
-      </div>
-
-      <div className={styles.subContent}>
-        <BlogPostLink heading="Recommended" posts={rankedPosts} />
-        <BlogTagLink heading="Categories" tags={tags} />
-      </div>
-    </div>
-  )
-}
-
-export default RenderPostsBeforeDate
